feat: derive the PDF name to be exported from the MD file name

+++ cmd/export.go
@@ -24,7 +24,10 @@ package cmd
 import (
 	"fmt"
 	"os"
+	"path/filepath"
+	"strings"
 
+	"github.com/Songmu/prompter"
 	"github.com/fatih/color"
 	"github.com/k1LoW/deck"
 	"github.com/k1LoW/deck/md"
@@ -47,7 +50,13 @@ var exportCmd = &cobra.Command{
 				// Deprecated
 				presentationID = f
 				cmd.Println(color.YellowString("WARNING: The argument is deprecated. Use --presentation-id instead."))
-				// return err
+			} else if out == "" {
+				// If no output destination is specified, output to a file name based on the MD file name.
+				// We could also consider generating file names from presentation names or titles, but from a
+				// local file management perspective, it is more intuitive to use the same file name.
+				// Furthermore, if we use presentation names, we need to sanitize the string to make filesystem safe.
+				// So, following the MD file name is a good default.
+				out = strings.TrimSuffix(filepath.Base(f), filepath.Ext(f)) + ".pdf"
 			}
 			if presentationID == "" && markdownData.Frontmatter != nil && markdownData.Frontmatter.PresentationID != "" {
 				presentationID = markdownData.Frontmatter.PresentationID
@@ -56,6 +65,11 @@ var exportCmd = &cobra.Command{
 		if presentationID == "" {
 			return fmt.Errorf("presentation ID is required. Use --presentation-id or set it in the frontmatter of the markdown file")
 		}
+		if out == "" {
+			// If the presentationID is passed as an argument (not recommended), "deck.pdf" will be used as a default output name.
+			out = "deck.pdf"
+		}
+
 		opts := []deck.Option{
 			deck.WithProfile(profile),
 			deck.WithPresentationID(presentationID),
@@ -63,6 +77,12 @@ var exportCmd = &cobra.Command{
 		d, err := deck.New(ctx, opts...)
 		if err != nil {
 			return err
+		}
+		if _, err = os.Stat(out); err == nil {
+			if !prompter.YN(fmt.Sprintf("%q already exists. Do you want to overwrite it?", out), false) {
+				fmt.Fprintln(os.Stderr, "The export has been canceled.")
+				return nil
+			}
 		}
 		f, err := os.Create(out)
 		if err != nil {
@@ -79,5 +99,5 @@ var exportCmd = &cobra.Command{
 func init() {
 	rootCmd.AddCommand(exportCmd)
 	exportCmd.Flags().StringVarP(&presentationID, "presentation-id", "i", "", "Google Slides presentation ID")
-	exportCmd.Flags().StringVarP(&out, "out", "o", "deck.pdf", "output file")
+	exportCmd.Flags().StringVarP(&out, "out", "o", "", `output file (default: follow the md file name, or "deck.pdf")`)
 }
